+++ src/python/pants/backend/experimental/java/register.py
@@ -1,7 +1,7 @@
 # Copyright 2021 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.backend.java import classpath, tailor
+from pants.backend.java import tailor
 from pants.backend.java.compile import javac
 from pants.backend.java.dependency_inference import (
     import_parser,
@@ -20,7 +20,7 @@ from pants.backend.java.target_types import (
 )
 from pants.backend.java.target_types import rules as target_types_rules
 from pants.backend.java.test import junit
-from pants.jvm import jdk_rules
+from pants.jvm import classpath, jdk_rules
 from pants.jvm import util_rules as jvm_util_rules
 from pants.jvm.goals import coursier
 from pants.jvm.resolve import coursier_fetch, coursier_setup
