+++ src/python/pants/backend/java/package/deploy_jar.py
@@ -7,7 +7,6 @@ import textwrap
 from dataclasses import dataclass
 from pathlib import PurePath
 
-from pants.backend.java.classpath import Classpath
 from pants.backend.java.target_types import JvmMainClassName
 from pants.core.goals.package import (
     BuiltPackage,
@@ -22,6 +21,7 @@ from pants.engine.process import BashBinary, Process, ProcessResult
 from pants.engine.rules import Get, collect_rules, rule
 from pants.engine.target import Dependencies, DependenciesRequest
 from pants.engine.unions import UnionRule
+from pants.jvm.classpath import Classpath
 
 logger = logging.getLogger(__name__)
